updated test cases

+++ Tests/TimedBin_Tests/TimedBin_Tests.h
@@ -50,13 +50,13 @@ test(TimedBinTest, shouldComputeTotals)
 
    loadTimedBinTestEntries(&bin, microsForEntries, numEntries);
 
-   assertEqual(5u, bin.getCount());
-   assertEqual(5u, bin.getBinCount(0));
-   assertEqual(0u, bin.getBinCount(1));
-   assertEqual(0u, bin.getBinCount(2));
-   assertEqual(0u, bin.getBinCount(3));
-   assertEqual(0u, bin.getBinCount(4));
-   assertEqual(0u, bin.getBinCount(-1));
+   assertEqual((uint16_t)5, bin.getCount());
+   assertEqual((uint16_t)5, bin.getBinCount(0));
+   assertEqual((uint16_t)0, bin.getBinCount(1));
+   assertEqual((uint16_t)0, bin.getBinCount(2));
+   assertEqual((uint16_t)0, bin.getBinCount(3));
+   assertEqual((uint16_t)0, bin.getBinCount(4));
+   assertEqual((uint16_t)0, bin.getBinCount(-1));
 
 }
 
@@ -81,13 +81,13 @@ test(TimedBinTest, shouldUtilizeMultipleBins)
 
    loadTimedBinTestEntries(&bin, microsForEntries, numEntries);
 
-   assertEqual(5u, bin.getCount());
-   assertEqual(1u, bin.getBinCount(0));
-   assertEqual(1u, bin.getBinCount(1));
-   assertEqual(1u, bin.getBinCount(2));
-   assertEqual(1u, bin.getBinCount(3));
-   assertEqual(1u, bin.getBinCount(4));
-   assertEqual(0u, bin.getBinCount(-1));
+   assertEqual((uint16_t)5, bin.getCount());
+   assertEqual((uint16_t)1, bin.getBinCount(0));
+   assertEqual((uint16_t)1, bin.getBinCount(1));
+   assertEqual((uint16_t)1, bin.getBinCount(2));
+   assertEqual((uint16_t)1, bin.getBinCount(3));
+   assertEqual((uint16_t)1, bin.getBinCount(4));
+   assertEqual((uint16_t)0, bin.getBinCount(-1));
 }
 
 test(TimedBinTest, shouldAdvanceBins)
@@ -108,13 +108,13 @@ test(TimedBinTest, shouldAdvanceBins)
 
    loadTimedBinTestEntries(&bin, microsForEntries, numEntries);
 
-   assertEqual(2u, bin.getCount());
-   assertEqual(1u, bin.getBinCount(0));
-   assertEqual(0u, bin.getBinCount(1));
-   assertEqual(0u, bin.getBinCount(2));
-   assertEqual(0u, bin.getBinCount(3));
-   assertEqual(1u, bin.getBinCount(4));
-   assertEqual(0u, bin.getBinCount(-1));
+   assertEqual((uint16_t)2, bin.getCount());
+   assertEqual((uint16_t)1, bin.getBinCount(0));
+   assertEqual((uint16_t)0, bin.getBinCount(1));
+   assertEqual((uint16_t)0, bin.getBinCount(2));
+   assertEqual((uint16_t)0, bin.getBinCount(3));
+   assertEqual((uint16_t)1, bin.getBinCount(4));
+   assertEqual((uint16_t)0, bin.getBinCount(-1));
 }
 
 // should zero out bins as they transition
@@ -140,27 +140,27 @@ test(TimedBinTest, shouldZeroOutBins)
    // first fill each bin
    loadTimedBinTestEntries(&bin, microsForEntries, numEntries);
 
-   assertEqual(5u, bin.getCount());
-   assertEqual(1u, bin.getBinCount(0));
-   assertEqual(1u, bin.getBinCount(1));
-   assertEqual(1u, bin.getBinCount(2));
-   assertEqual(1u, bin.getBinCount(3));
-   assertEqual(1u, bin.getBinCount(4));
-   assertEqual(0u, bin.getBinCount(-1));
+   assertEqual((uint16_t)5, bin.getCount());
+   assertEqual((uint16_t)1, bin.getBinCount(0));
+   assertEqual((uint16_t)1, bin.getBinCount(1));
+   assertEqual((uint16_t)1, bin.getBinCount(2));
+   assertEqual((uint16_t)1, bin.getBinCount(3));
+   assertEqual((uint16_t)1, bin.getBinCount(4));
+   assertEqual((uint16_t)0, bin.getBinCount(-1));
 
    // advance long enough to expire the first bin
    timedBinTestMockMicros = (DURATION_MS + 1) * 1000;
-   assertEqual(5u, bin.getCount());
-   assertEqual(0u, bin.getBinCount(0));
-   assertEqual(1u, bin.getBinCount(1));
-   assertEqual(1u, bin.getBinCount(2));
-   assertEqual(1u, bin.getBinCount(3));
-   assertEqual(1u, bin.getBinCount(4));
-   assertEqual(1u, bin.getBinCount(-1));
+   assertEqual((uint16_t)5, bin.getCount());
+   assertEqual((uint16_t)0, bin.getBinCount(0));
+   assertEqual((uint16_t)1, bin.getBinCount(1));
+   assertEqual((uint16_t)1, bin.getBinCount(2));
+   assertEqual((uint16_t)1, bin.getBinCount(3));
+   assertEqual((uint16_t)1, bin.getBinCount(4));
+   assertEqual((uint16_t)1, bin.getBinCount(-1));
 
    // expire all bins - should then be zero
    timedBinTestMockMicros = (2 * DURATION_MS + 1) * 1000;
-   assertEqual(0u, bin.getCount());
+   assertEqual((uint16_t)0, bin.getCount());
 }
 
 // should smoothly transition between bins as time advances
@@ -186,30 +186,30 @@ test(TimedBinTest, shouldSmoothlyTransitionBins)
    // fill bins
    loadTimedBinTestEntries(&bin, microsForEntries, numEntries);
 
-   assertEqual(5u, bin.getCount());
-   assertEqual(5u, bin.getBinCount(0));
-   assertEqual(0u, bin.getBinCount(1));
-   assertEqual(0u, bin.getBinCount(2));
-   assertEqual(0u, bin.getBinCount(3));
-   assertEqual(0u, bin.getBinCount(4));
-   assertEqual(0u, bin.getBinCount(-1));
+   assertEqual((uint16_t)5, bin.getCount());
+   assertEqual((uint16_t)5, bin.getBinCount(0));
+   assertEqual((uint16_t)0, bin.getBinCount(1));
+   assertEqual((uint16_t)0, bin.getBinCount(2));
+   assertEqual((uint16_t)0, bin.getBinCount(3));
+   assertEqual((uint16_t)0, bin.getBinCount(4));
+   assertEqual((uint16_t)0, bin.getBinCount(-1));
 
    // gradually expire the first bin
    constexpr auto BIN_DURATION_MS = DURATION_MS / NUM_BINS;
    timedBinTestMockMicros = (DURATION_MS + 0.21*BIN_DURATION_MS) * 1000;
-   assertEqual(4u, bin.getCount());
+   assertEqual((uint16_t)4, bin.getCount());
 
    timedBinTestMockMicros = (DURATION_MS + 0.41 * BIN_DURATION_MS) * 1000;
-   assertEqual(3u, bin.getCount());
+   assertEqual((uint16_t)3, bin.getCount());
 
    timedBinTestMockMicros = (DURATION_MS + 0.61 * BIN_DURATION_MS) * 1000;
-   assertEqual(2u, bin.getCount());
+   assertEqual((uint16_t)2, bin.getCount());
 
    timedBinTestMockMicros = (DURATION_MS + 0.81 * BIN_DURATION_MS) * 1000;
-   assertEqual(1u, bin.getCount());
+   assertEqual((uint16_t)1, bin.getCount());
 
    timedBinTestMockMicros = (DURATION_MS + 1.01 * BIN_DURATION_MS) * 1000;
-   assertEqual(0u, bin.getCount());
+   assertEqual((uint16_t)0, bin.getCount());
 }
 
 test(TimedBinTest, shouldHonorBeginAsNewStartTime)
@@ -227,12 +227,12 @@ test(TimedBinTest, shouldHonorBeginAsNewStartTime)
    bin.begin();
    bin.add();
 
-   assertEqual(2u, bin.getCount());
-   assertEqual(2u, bin.getBinCount(0));
-   assertEqual(0u, bin.getBinCount(1));
-   assertEqual(0u, bin.getBinCount(2));
-   assertEqual(0u, bin.getBinCount(3));
-   assertEqual(0u, bin.getBinCount(4));
+   assertEqual((uint16_t)2, bin.getCount());
+   assertEqual((uint16_t)2, bin.getBinCount(0));
+   assertEqual((uint16_t)0, bin.getBinCount(1));
+   assertEqual((uint16_t)0, bin.getBinCount(2));
+   assertEqual((uint16_t)0, bin.getBinCount(3));
+   assertEqual((uint16_t)0, bin.getBinCount(4));
 }
 
 } // namespace TimedBinTests
